Changes related to database adaptation All DAOs and domains adapted to new database types; Methods parameters and returns adapted; Minor updates.

+++ app/src/main/java/com/enoughspam/step/database/dao/related/CountryDAO.java
@@ -40,7 +40,7 @@ public class CountryDAO {
         );
     }
 
-    public Country findById(long id) {
+    public Country findById(int id) {
         String sql = "select * from country where id = " + id;
         Cursor cursor = getSqLiteDatabase().rawQuery(sql, null);
